risoluzione di bug di eliminazione e visualizzazione

+++ budget.sql
@@ -1,7 +1,7 @@
 CREATE TABLE Budget (
     numero_spesa INTEGER PRIMARY KEY AUTOINCREMENT,
     categoria TEXT NOT NULL,
-    importo INTEGER NOT NULL,
+    importo DOUBLE NOT NULL,
     data TEXT NOT NULL
 )
